requirements.txt
# Flask
Flask==3.0.0
# flask-api-key==0.2.12
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0

# Utils
passlib==1.7.4
PyJWT==2.8.0
importlib-metadata==7.0.1
python-dotenv==1.0.0
setuptools

# Data Scientists
numpy==1.26.2
polars==0.20.0
xlsx2csv==0.8.2
# openpyxl==3.1.2
# pandas==2.1.4

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Docs
flasgger==0.9.7.1
marshmallow==3.20.1
apispec==6.3.1

# Tests
coverage==7.3.4
requests==2.31.0

# Machine learning
scikit-learn==1.3.2
scipy==1.11.4

# Face recognition
cmake==3.28.1
dlib==19.24.2
face-recognition==1.3.0

# Cache
redis==5.0.1
flask-redis==0.4.0